+++ include/kmeans_training_data.h
@@ -9,9 +9,9 @@
 
 void KmeansDataAndLabels(
     const std::string& directory, int num_centers, int& width, int& height,
-    float** training_data, int* training_data_dim1, int* training_data_dim2,
-    float** training_labels, int* training_labels_dim1,
-    int* training_labels_dim2, float** average_img, int* average_dim1,
+    double** training_data, int* training_data_dim1, int* training_data_dim2,
+    double** training_labels, int* training_labels_dim1,
+    int* training_labels_dim2, double** average_img, int* average_dim1,
     int* average_dim2, int* average_dim3, std::vector<int>& indices,
     std::vector<int>& order, std::vector<glm::vec2>& centers,
     std::vector<int>& labels, std::vector<int>& batch_sizes);
